// packages/core/src/rpe/candidate-generator-types.ts
import { LLMUsage } from '../llm/index.js';
import { RPECandidate, RPECandidateRef } from './rpe-candidate/index.js';
import { 
  RPECandidateGeneratorChange,
} from './rpe-candidate-generator/index.js';

/**
 * Candidate prompt generated by the RPE prompt generator.
 * @category Reflective Prompt Evolution
 */
export interface PromptGeneratorOutputCandidate {
  /**
   * New candidate.
   */
  candidate: RPECandidate;

  /**
   * Changes made to the prompt.
   */
  changes: RPECandidateGeneratorChange[];

  /**
   * Usage of the model.
   */
  usage?: LLMUsage;
}

/**
 * Output for the RPE prompt generator.
 * @category Reflective Prompt Evolution
 */
export interface PromptGeneratorOutput {
  /**
   * New candidate candidates.
   */
  candidates: PromptGeneratorOutputCandidate[];
}

/**
 * Candidate prompt for the RPE prompt generator stored in the state.
 * @category Reflective Prompt Evolution
 */
export interface PromptGeneratorCandidate {
  /**
   * Reference to the new candidate.
   */
  candidateRef: RPECandidateRef;

  /**
   * Changes made to the prompt.
   */
  changes: RPECandidateGeneratorChange[];

  /**
   * Usage of the model.
   */
  usage?: LLMUsage;
} 